% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASC.R
\name{adj2vector}
\alias{adj2vector}
\title{Transform the adjacency matrix to a vector}
\usage{
adj2vector(Adjacency, n)
}
\arguments{
\item{Adjacency}{the adjacency matrix of factor}

\item{n}{number of samples}
}
\value{
\code{w} the vector of the adjacency matrix
}
\description{
Transform the adjacency matrix to a vector
}
\examples{
W <- matrix(c(0,1,0,0,1,0,0,0,0,0,0,0,0,0,0,0),nrow=4)
w <- adj2vector(W,4)

}
\author{
Haidong Yi, Ayush T. Raman
}
