% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASC-package.R, R/DASC.R
\docType{package}
\name{DASC}
\alias{DASC}
\alias{DASC-package}
\alias{DASC}
\title{Detecting hidden batch factors through data adaptive shrinkage 
and clustering (DASC)}
\usage{
DASC(edata, pdata, factor, method = "ama", type = 3, lambda, rank, nrun,
  spanning = FALSE, annotation)
}
\arguments{
\item{edata}{the normalized target matrix, a data.frame The row is gene,
the column is sample}

\item{pdata}{Phenotypic data summarizes information about the samples}

\item{factor}{A factor vector which controls the convex clustering}

\item{method}{Algorithm to use: 'admm' or 'ama'}

\item{type}{An integer indicating the norm used: 1 = 1-norm 2 = 2-norm
3 = 2-norm^2}

\item{lambda}{A double number A regularization parameter in the convex
optimization}

\item{rank}{integer sequence}

\item{nrun}{the iteration numbers of Semi-NMF}

\item{spanning}{parameter is assigned as false}

\item{annotation}{An annotation of the dataset}
}
\value{
outputs the result of semi-NMF. It classifies each sample to its
batch factor.
}
\description{
Detecting hidden batch factors through data adaptive shrinkage 
and clustering (DASC)

Batch factor detection via DASC (Data-adaptive Shrinkage and Clustering-DASC)
}
\details{
The \code{DASC} function is the main function of our algorithm DASC
(Data-adaptive Shrinkage and Clustering-DASC) package. The DASC includes 
two main steps
\itemize{ \item Data-adaptive shrinkage using convex clustering shrinkage
(Implemented by convex optimization.);
\item Extract batch factors using matrix factorization.
}
}
\examples{
library(NMF)
library(cvxclustr)
library(Biobase)
dat <- data.frame(matrix(rnbinom(n=200, mu=100, size=1/0.5), ncol=4))
pdat <- data.frame(sample = colnames(dat), type = c(rep('A',2), rep('B',2)))
rownames(pdat) <- colnames(dat)
res <- DASC(edata=dat, pdata=pdat, factor=pdat$type, method='ama', type=3,
lambda = 1, rank = 2, nrun = 50, spanning = FALSE,
annotation='simulated dataset')

}
\seealso{
\code{\link[cvxclustr]{cvxclust_path_ama}} and
\code{\link[cvxclustr]{cvxclust_path_admm}} for the detailed algorithm
}
\author{
Haidong Yi, Ayush T. Raman

Haidong Yi, Ayush T. Raman
}
\keyword{package}
