% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets}
\alias{datasets}
\title{Metadata information of 26 samples}
\format{A data frame with 26 rows and 4 variables:}
\source{
\url{http://www.pnas.org/content/111/48/17224.abstract}
}
\usage{
datasets
}
\description{
A dataset containing information about tissues used, sequencer (along with lane) and species.
}
\keyword{datasets}
