% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_batch.R
\name{get_class}
\alias{get_class}
\title{Spanning tree from adjacency matrix}
\usage{
get_class(ADJ)
}
\arguments{
\item{ADJ}{the adjacency matrix of the factor}
}
\value{
\code{ADJ} the spaning tree of the adjacency matrix
}
\description{
Spanning tree from adjacency matrix
}
\examples{
W <- matrix(c(0,1,1,1,1,0,1,1,1,1,0,1,1,1,1,0), nrow=4)
get_class(W)

}
\author{
Haidong Yi, Ayush T. Raman
}
