% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASC.R
\name{get_father}
\alias{get_father}
\title{Representing node in this subtype}
\usage{
get_father(v, X)
}
\arguments{
\item{v}{the index of the node}

\item{X}{the saved vector with the information of the parent of every node}
}
\value{
\code{r} the parent index of the node
}
\description{
Representing node in this subtype
}
\examples{
nodes <- c(2,3,4,4)
get_father(2, nodes)

}
\author{
Haidong Yi, Ayush T. Raman
}
