% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexBatch.R
\name{getclass}
\alias{getclass}
\title{Spanning tree from adjacency matrix}
\usage{
getclass(ADJ)
}
\arguments{
\item{ADJ}{the adjacency matrix of the factor}
}
\value{
\code{ADJ} the spaning tree of the adjacency matrix
}
\description{
Spanning tree from adjacency matrix
}
\examples{
W <- matrix(c(0,1,1,1,1,0,1,1,1,1,0,1,1,1,1,0), nrow=4)
getclass(W)

}
\author{
Haidong Yi, Ayush T. Raman
}
