% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexBatch.R
\name{getFather}
\alias{getFather}
\title{Representing node in this subtype}
\usage{
getFather(v, X)
}
\arguments{
\item{v}{the index of the node}

\item{X}{the saved vector with the information of the parent of every node}
}
\value{
\code{r} the parent index of the node
}
\description{
Representing node in this subtype
}
\author{
Haidong Yi, Ayush T. Raman
}
