% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ini_SemiNMF.R
\name{Ini_SemiNMF}
\alias{Ini_SemiNMF}
\title{Initialization of the Semi-NMF}
\usage{
Ini_SemiNMF(target, rank)
}
\arguments{
\item{target}{gene expression matrix}

\item{rank}{rank used to decompose the target}
}
\value{
The initial objective of class: NMFfit
}
\description{
Initialization of the Semi-NMF
}
\author{
Haidong Yi, Ayush T. Raman
}
