% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loss_Fro.R
\name{Loss_Fro}
\alias{Loss_Fro}
\title{Get the error of Semi-NMF using frobenius norm}
\usage{
Loss_Fro(model, target)
}
\arguments{
\item{model}{Object of class: NMFfit}

\item{target}{gene expression matrix}
}
\value{
The result of semi-NMF for the current iteration
}
\description{
Get the error of Semi-NMF using frobenius norm
}
\details{
This is a customerized function defined in terms of \code{\link[NMF]{nmf}}.
For more information, please go through the NMF vignette
\url{https://cran.r-project.org/web/packages/NMF/vignettes/NMF-vignette.pdf}
}
\author{
Haidong Yi, Ayush T. Raman
}
