% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASC.R
\name{merge}
\alias{merge}
\title{Combine two trees into one}
\usage{
merge(x, y, X)
}
\arguments{
\item{x}{the index of the node}

\item{y}{the index of the node}

\item{X}{the saved vector with the information of the parent of every node}
}
\value{
\code{X} A updated X vector with updates on father of every node
}
\description{
Combine two trees into one
}
\details{
merge combines two trees into one by attaching the root of one to the root 
of the other. It is a step of the Disjoint-set Algorithm/Union-find 
Algorithm invented by Bernard A. Galler and Michael J. Fischer. The concrete 
details can be found at 
\url{https://en.wikipedia.org/wiki/Disjoint-set_data_structure}
}
\author{
Haidong Yi, Ayush T. Raman
}
