% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.algorithm.R
\name{my.algorithm}
\alias{my.algorithm}
\title{Main function of semi-NMF}
\usage{
my.algorithm(target, model, iternum = 100)
}
\arguments{
\item{target}{gene expression matrix}

\item{model}{Object of class: NMFfit}

\item{iternum}{Number of iterations}
}
\value{
\code{model} the result from the semi-NMF algorithm
}
\description{
Main function of semi-NMF
}
\author{
Haidong Yi, Ayush T. Raman
}
