% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_objective_function.R
\name{my_objective_function}
\alias{my_objective_function}
\title{Get the error of Semi-NMF}
\usage{
my_objective_function(model, target)
}
\arguments{
\item{model}{Object of class: NMFfit}

\item{target}{gene expression matrix}
}
\value{
The result of semi-NMF for the current iteration
}
\description{
Get the error of Semi-NMF
}
\details{
This is a customerized function defined in terms of \code{\link[NMF]{nmf}}. 
For more information, please go through the NMF vignette 
\url{https://cran.r-project.org/web/packages/NMF/vignettes/NMF-vignette.pdf}
}
\author{
Haidong Yi, Ayush T. Raman
}
