% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_seeding_method.R
\name{my_seeding_method}
\alias{my_seeding_method}
\title{Initialization of the semi-NMF}
\usage{
my_seeding_method(model, target)
}
\arguments{
\item{model}{Object of class: NMFfit}

\item{target}{gene expression matrix}
}
\value{
\code{model} The initial objective of class: NMFfit
}
\description{
Initialization of the semi-NMF
}
\author{
Haidong Yi, Ayush T. Raman
}
