% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Semi_NMF.R
\name{pred.consen}
\alias{pred.consen}
\title{Predict the classification using consensus matrices}
\usage{
pred.consen(consen, rank)
}
\arguments{
\item{consen}{a consensus matrix}

\item{rank}{rank number using to cut the tree}
}
\value{
\code{} a factor vector representing the category of samples
}
\description{
Predict the classification using consensus matrices
}
\author{
Haidong Yi
}