% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Semi_NMF.R
\name{pred.nmf}
\alias{pred.nmf}
\title{Predict function of semi-NMF to determine which category are the samples assigned}
\usage{
pred.nmf(x)
}
\arguments{
\item{x}{transpose of the coefficient matrix}
}
\value{
\code{} a factor vector representing the classification
}
\description{
Predict function of semi-NMF to determine which category are the samples assigned
}
\author{
Haidong Yi
}