% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rawCounts}
\alias{rawCounts}
\title{Stanford Filtered Raw Counts dataset.}
\format{A matrix with 14744 rows and 26 variables:}
\source{
\url{https://f1000research.com/articles/4-121/v1}
}
\usage{
rawCounts
}
\description{
A gene expression dataset containing values for 14744 genes across 26 different samples
}
\keyword{datasets}
