% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Semi_NMF.R
\name{Semi_NMF}
\alias{Semi_NMF}
\title{Main function of semi-NMF}
\usage{
Semi_NMF(target, rank, iternum = 100, tol = 0.01)
}
\arguments{
\item{target}{gene expression matrix}

\item{rank}{rank used to decompose target}

\item{iternum}{Number of iterations}

\item{tol}{tolerance for convergence}
}
\value{
\code{model} the result from the semi-NMF algorithm
}
\description{
Main function of semi-NMF
}
\author{
Haidong Yi, Ayush T. Raman
}
