% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Semi_NMF.R
\name{sNMF}
\alias{sNMF}
\title{Parallel implement Semi-NMF for different ranks and times}
\usage{
sNMF(target, rank, nrun = 30)
}
\arguments{
\item{target}{target matrix}

\item{rank}{rank used to decompose the target}

\item{nrun}{times to run Semi-NMF}
}
\value{
\code{snmf} a list containing consensus matrix, classification and dispersion
}
\description{
Parallel implement Semi-NMF for different ranks and times
}
\author{
Haidong Yi
}