% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_ADJ.R
\name{trans_ADJ}
\alias{trans_ADJ}
\title{Outputs Adjacency matrix from the factor vector}
\usage{
trans_ADJ(col_data)
}
\arguments{
\item{col_data}{A factor vector}
}
\value{
\code{adjacency} Adjacency matrix of \code{col_data}
}
\description{
Outputs Adjacency matrix from the factor vector
}
\examples{
batch.factor <- c(rep('human',13),rep('mouse',13))
batch.factor <- as.factor(batch.factor)
adj <- trans_Laplace(batch.factor)

}
\author{
Haidong Yi, Ayush T. Raman
}
