% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_Laplace.R
\name{trans_Laplace}
\alias{trans_Laplace}
\title{Get Laplace matrix from factor vector}
\usage{
trans_Laplace(col_data)
}
\arguments{
\item{col_data}{A factor vector}
}
\value{
\code{adjacency} The Laplace matrix of \code{col_data}
}
\description{
Get Laplace matrix from factor vector
}
\examples{
batch.factor <- c(rep('human',13),rep('mouse',13))
batch.factor <- as.factor(batch.factor)
adj <- trans_Laplace(batch.factor)

}
\author{
Haidong Yi, Ayush T. Raman
}
