% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_G.R
\name{update_G}
\alias{update_G}
\title{Update G in Semi-NMF}
\usage{
update_G(X, mf, mg)
}
\arguments{
\item{X}{Data expression matrix need to be factorized}

\item{mf}{The basis matrix}

\item{mg}{The co-efficient matrix}
}
\value{
\code{G} The basis matrix
}
\description{
Update G in Semi-NMF
}
\details{
By definition, G is a graph adjacency matrix. The \code{update_G} updates G 
after every iteration.
}
\examples{
X <- matrix(1:12,nrow=4)
mf <- matrix(1:8,nrow=4)
mg <- matrix(1:6,ncol=2)
mg <- update_G(X,mf,mg)
}
\author{
Haidong Yi, Ayush T. Raman
}
