% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateG.R
\name{updateG}
\alias{updateG}
\title{Update G in Semi-NMF}
\usage{
updateG(X, mf, mg)
}
\arguments{
\item{X}{Data expression matrix need to be factorized}

\item{mf}{The basis matrix}

\item{mg}{The co-efficient matrix}
}
\value{
\code{G} The basis matrix
}
\description{
Update G in Semi-NMF
}
\examples{
X <- matrix(1:12,nrow=4)
mf <- matrix(1:8,nrow=4)
mg <- matrix(1:6,ncol=2)
mg <- updateG(X,mf,mg)
}
\author{
Haidong Yi, Ayush T. Raman
}
