% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexBatch.R
\name{W_toVector}
\alias{W_toVector}
\title{Transform the adjacency matrix to a vector}
\usage{
W_toVector(Adjacency, n)
}
\arguments{
\item{Adjacency}{the adjacency matrix of factor}

\item{n}{number of samples}
}
\value{
\code{w} the vector of the adjacency matrix
}
\description{
Transform the adjacency matrix to a vector
}
\examples{
W <- matrix(c(0,1,0,0,1,0,0,0,0,0,0,0,0,0,0,0),nrow=4)
w <- W_toVector(W,4)
}
\author{
Haidong Yi, Ayush T. Raman
}
