% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASC-package.R
\docType{package}
\name{DASC}
\alias{DASC}
\alias{DASC-package}
\title{Detecting hidden batch factors through data adaptive shrinkage and clustering (DASC)}
\description{
Detecting hidden batch factors through data adaptive shrinkage and clustering (DASC)
}
\author{
Haidong Yi, Ayush T. Raman
}
