% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ini_SemiNMF.R
\name{Ini_SemiNMF}
\alias{Ini_SemiNMF}
\title{Initialization of the semi-NMF}
\usage{
Ini_SemiNMF(model, target)
}
\arguments{
\item{model}{Object of class: NMFfit}

\item{target}{gene expression matrix}
}
\value{
\code{model} The initial objective of class: NMFfit
}
\description{
Initialization of the semi-NMF
}
\author{
Haidong Yi, Ayush T. Raman
}
