% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexBatch.R
\name{merge}
\alias{merge}
\title{Merge two nodes}
\usage{
merge(x, y, X)
}
\arguments{
\item{x}{the index of the node}

\item{y}{the index of the node}

\item{X}{the saved vector with the information of the parent of every node}
}
\value{
\code{X} A updated X vector with updates on father of every node
}
\description{
Merge two nodes
}
\author{
Haidong Yi, Ayush T. Raman
}
