% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASC.R
\name{merge}
\alias{merge}
\title{Combine two trees into one}
\usage{
merge(x, y, X)
}
\arguments{
\item{x}{the index of the node}

\item{y}{the index of the node}

\item{X}{the saved vector with the information of the parent of every node}
}
\value{
\code{X} A updated X vector with updates on father of every node
}
\description{
Combine two trees into one
}
\details{
During the traversal of the graph matrix, merge function joins two 
disjoint sets into a single subset. It is a union step of Disjoint-set 
algorithm by Bernard A. Galler and Michael J. Fischer. For further details, 
please refer to: 
\url{https://en.wikipedia.org/wiki/Disjoint-set_data_structure}
}
\author{
Haidong Yi, Ayush T. Raman
}
