% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfordData-data.R
\docType{data}
\name{stanfordData}
\alias{stanfordData}
\title{Stanford Dataset}
\format{A list with two components.
\describe{
\item{rawCounts}{ is a DataFrame object where each column represents a 
sample and each row represents a gene.}
\item{metadata}{ is a metadata information about each samples}
}}
\source{
\url{https://f1000research.com/articles/4-121/v1}
}
\description{
Stanford Dataset
}
\references{
1. Gilad Y and Mizrahi-Man O. A reanalysis of mouse ENCODE comparative gene 
expression data. F1000Research (2015)

2. in S, Lin Y, Nery JR, et al.: Comparison of the transcriptional 
landscapes between human and mouse tissues. Proc Natl Acad Sci USA (2014)
}
\keyword{datasets}
